MAR-3160. removed Footer and change FooterV2 to Footer
