+++ client/components/core/Footer/FooterStaticNavigations.vue
