+++ client/layouts/error-layout.vue
@@ -8,7 +8,7 @@
 
 <script>
 import HeaderV2 from '@/components/core/HeaderV2'
-import Footer from '@/components/core/FooterV2/Footer'
+import Footer from '@/components/core/Footer/Footer'
 
 export default {
   name: 'ErrorLayout',
